cleaner/requirements.txt
pandas>=2.2
duckdb>=1.0
minio>=7.2.7
pika>=1.3.2
python-dotenv>=1.0